♻️ refactor: refactor the openai initial code to the createBizOpenAI
