+++ src/app/api/openai/stt/route.ts
@@ -1,55 +1,19 @@
 import { OpenAISTTPayload } from '@lobehub/tts';
 import { createOpenaiAudioTranscriptions } from '@lobehub/tts/server';
-import OpenAI from 'openai';
 
-import { getServerConfig } from '@/config/server';
-import { getOpenAIAuthFromRequest } from '@/const/fetch';
-import { ChatErrorType, ErrorType } from '@/types/fetch';
-
-import { checkAuth } from '../../auth';
-import { createAzureOpenai } from '../createAzureOpenai';
-import { createOpenai } from '../createOpenai';
-import { createErrorResponse } from '../errorResponse';
+import { createBizOpenAI } from '../createBizOpenAI';
 
 export const runtime = 'edge';
 
 export const POST = async (req: Request) => {
   const payload = (await req.json()) as OpenAISTTPayload;
 
-  const { apiKey, accessCode, endpoint, useAzure, apiVersion } = getOpenAIAuthFromRequest(req);
+  const openaiOrErrResponse = createBizOpenAI(req, payload.options.model);
 
-  const result = checkAuth({ accessCode, apiKey });
+  // if resOrOpenAI is a Response, it means there is an error,just return it
+  if (openaiOrErrResponse instanceof Response) return openaiOrErrResponse;
 
-  if (!result.auth) {
-    return createErrorResponse(result.error as ErrorType);
-  }
-
-  let openai: OpenAI;
-
-  const { USE_AZURE_OPENAI } = getServerConfig();
-  const useAzureOpenAI = useAzure || USE_AZURE_OPENAI;
-
-  try {
-    if (useAzureOpenAI) {
-      openai = createAzureOpenai({
-        apiVersion,
-        endpoint,
-        model: payload.options.model,
-        userApiKey: apiKey,
-      });
-    } else {
-      openai = createOpenai(apiKey, endpoint);
-    }
-  } catch (error) {
-    if ((error as Error).cause === ChatErrorType.NoAPIKey) {
-      return createErrorResponse(ChatErrorType.NoAPIKey);
-    }
-
-    console.error(error); // log error to trace it
-    return createErrorResponse(ChatErrorType.InternalServerError);
-  }
-
-  const res = await createOpenaiAudioTranscriptions({ openai, payload });
+  const res = await createOpenaiAudioTranscriptions({ openai: openaiOrErrResponse, payload });
 
   return new Response(JSON.stringify(res), {
     headers: {
